FastApi/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.121
botocore==1.29.121
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
fastapi==0.95.1
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
jwt==1.3.1
loguru==0.7.0
multidict==6.0.4
numpy==1.24.3
openai==0.27.4
passlib==1.7.4
pinecone-client==2.2.1
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.29.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
yarl==1.9.2
